feat: 개인 채팅 기능 완료(#71_personal_chatting_function)

+++ TripFriend/app/src/main/java/com/test/tripfriend/dataclassmodel/PersonalChat.kt
@@ -15,7 +15,7 @@ data class PersonalChatting(
     val personalChatWriterEmail: String,           //채팅 작성자 이메일
     val personalChatContent: String,               //채팅 내용
     val personalChatSendDateAndTime: String, //채팅 작성 날짜 및 시간
-    val personalChatSendTimeStamp: Int,   //채팅 저장 시기
+    val personalChatSendTimeStamp: Long,   //채팅 저장 시기
 )
 
 //서버에서 값을 가져오기 위한 생성자 없는 클래스
@@ -23,7 +23,7 @@ class PersonalChatting2 {
     val personalChatWriterEmail: String? = null           //채팅 작성자 이메일
     val personalChatContent: String? = null               //채팅 내용
     val personalChatSendDateAndTime: String? = null //채팅 작성 날짜 및 시간
-    val personalChatSendTimeStamp: Int? = null   //채팅 저장 시기
+    val personalChatSendTimeStamp: Long? = null   //채팅 저장 시기
 }
 
 //1:1 채팅방 리스트에 필요한 정보를 저장하고 사용하기 위한 클래스
